create app store folder and move files
