+++ src/Components/AppStore/AppStoreListing.jsx
@@ -17,7 +17,7 @@ const StoreData = [
     appName: 'vyzn',
     description: 'Perform environemental analysis on your model',
     image: 'https://www.vyzn.tech/wp-content/themes/stuiq-base/assets/images/logo.svg',
-    action: 'http://127.0.0.1:5501/',
+    action: 'http://127.0.0.1:5500/web/HostedAppSample.html',
   },
   // {
   //   appName: 'vyzn',
