+++ src/Containers/CadView.jsx
@@ -10,7 +10,7 @@ import Logo from '../Components/Logo'
 import NavPanel from '../Components/NavPanel'
 import SearchBar from '../Components/SearchBar'
 import SideDrawer from '../Components/SideDrawer/SideDrawer'
-import AppStoreSideDrawer from '../Components/SideDrawer/AppStoreSideDrawer'
+import AppStoreSideDrawer from '../Components/AppStore/AppStoreSideDrawer'
 import OperationsGroup from '../Components/OperationsGroup'
 import SnackBarMessage from '../Components/SnackbarMessage'
 import {hasValidUrlParams as urlHasCameraParams} from '../Components/CameraControl'
